requirements.txt
pillow
pytesseract
streamlit
requests